+++ apps/dv-admin/src/views/auth/signUp/SignUpView.vue
@@ -9,8 +9,6 @@
 	import { useRouter } from "vue-router";
 	import VueTurnstile from "vue-turnstile";
 	import { useSystemSettingsStore } from "@dv-admin/stores/systemSettings";
-	import policyFile from "@dv-admin/assets/docs/Privacy-Policy.pdf";
-	import termsFile from "@dv-admin/assets/docs/End-User-Agreement.pdf";
 	import { useI18n } from "vue-i18n";
 
 	const { isLoading } = storeToRefs(useAuthStore());
@@ -103,9 +101,9 @@
 		<ui-form-item>
 			<p class="checkbox-text">
 				{{ $t("By continuing registration you agree to") }} <br />
-				<ui-link :href="termsFile" target="_blank">{{ $t("Terms of Use") }}</ui-link>
+				<ui-link href="https://dv.net/files/End-User-Agreement.pdf" target="_blank">{{ $t("Terms of Use") }}</ui-link>
 				{{ $t("and") }}
-				<ui-link :href="policyFile" target="_blank">{{ $t("Privacy Policy") }}</ui-link>
+				<ui-link href="https://dv.net/files/Privacy-Policy.pdf" target="_blank">{{ $t("Privacy Policy") }}</ui-link>
 			</p>
 		</ui-form-item>